+++ NumberList/NumberList.h
@@ -1,5 +1,6 @@
 // Specification file for the NumberList class
 #pragma once
+#include <iostream>
 
 class NumberList
 {
@@ -7,7 +8,7 @@ private:
 	// Declare a structure for the list
 	struct ListNode
 	{
-		double value;           // The value in this node
+		int value;           // The value in this node
 		struct ListNode *next;  // To point to the next node
 	};
 
@@ -20,13 +21,14 @@ public:
 		head = nullptr;
 	}
 
+
 	// Destructor
 	~NumberList();
 
 	// Linked list operations
-	void appendNode(double);
-	void insertNode(double);
-	void deleteNode(double);
+	void appendNode(int);
+	void insertNode(int);
+	void deleteNode(int);
 	void displayList() const;
 };
